Add support for 2-dimensional scrolling with touch pads
Adds support for 2-dimensional scrolling, which is useful for touch pad
users. In order to facilitate this, the keyboard symbols that scroll
events map into were changed into the keypad up/down/left/right keys
(so now we have four symbols rather than two), and default bindings
were added for those symbols so that two-dimensional touch pad scrolling
works out-of-the-box.

Signed-off-by: Spaceship Operations <spaceshipoperations@gmail.com>